requirements.txt
pyspark==3.5.0
python-dotenv==1.0.0
numpy==1.26.2